+++ lib/phoenix_live_view/plug.ex
@@ -50,11 +50,11 @@ defmodule Phoenix.LiveView.Plug do
   end
 
   defp put_new_layout_from_router(conn, opts) do
-    cond do
-      live_link?(conn) -> Phoenix.Controller.put_layout(conn, false)
-      layout = opts[:layout] -> Phoenix.Controller.put_layout(conn, layout)
-      layout = opts[:inferred_layout] -> Phoenix.Controller.put_new_layout(conn, layout)
-      true -> conn
+    if live_link?(conn) do
+      Phoenix.Controller.put_layout(conn, false)
+    else
+      layout = opts[:layout] || conn.private[:phoenix_live_layout] || false
+      Phoenix.Controller.put_layout(conn, layout)
     end
   end