+++ test/support/router.ex
@@ -15,7 +15,7 @@ defmodule Phoenix.LiveViewTest.Router do
   end
 
   pipeline :bad_layout do
-    plug :put_layout, {UnknownView, :unknown_template}
+    plug :put_live_layout, {UnknownView, :unknown_template}
   end
 
   scope "/", Phoenix.LiveViewTest do